Updates with respect to the issue https://github.com/billh0420/ClaimGPT250203/issues/5
